export_stats_multi_year/requirements.txt
sqlalchemy
psycopg2
python-dotenv
pandas